Refactored how item references are handled for CdsListItem
Ownwership transfers are also made explicit. This change is about making
the interface easier to use for 99.9% of the cases, where the ownership
of an item is semantically transferred every time an item is pushed to
and popped from a list.

For the 0.1% of special cases, the caller can add/remove references to
items explicitely.